Backport: Add size to StaticArray + Unify SArray,SMatrix,SVector etc
This is two commits squashed together, with significant modifications to
backport to julia-0.5.

b1714d9 -----------------------------

* Unify SArray,SMatrix,SVector

After the recent refactor something that used to be an SMatrix became a
two dimensional SArray (or vice versa - who can keep track). That situation
seems unfortunate and unnecessary, so this is an attempt to unify them.
This is decently breaking, because it requires the SArray dimension tuple
to be e.g. `Tuple{1,2}` rather than `(1,2)`, but should allow us to avoid
quite a bit of duplication. These are just the minimal changes to get this
running. I would imagine, we could remove some further duplication in follow
up patches.

* Bring Scalar/MArray into the fold

c4008ff -----------------------------

WIP Add size to abstract StaticArray type (#134)

* Add size to abstract StaticArray type
* Update SUnitRange + bufix