Make `broadcast` signature compatible with `Base`.
A recent PR made some other signatures not define the zero-input version of certain functions where they are not defined in Base. In this case, this seems somewhat unnecessary since there is already a `broadcast(f, ::Number...)` method in `Base`.

The alternative is to define `broadcast(f, ::Number)` method here to remove the ambiguity.