SizeClass trait to control dispatch to small-sized algorithm specializations
The idea here is that you might want one algorithm for all matrices
smaller than a given size, and another for large matrices (for example,
dispatching to Base).  Rather than explicitly writing out methods to
dispatch to all small sizes, this allows the dispatch to be controlled
based on a size threshold.