Use promotion rather than typejoin to infer map/broadcast output eltype
This allows `Union`s as the output type for mixtures of `Null` and
numeric types, which can be more efficient than the `Any` you'd get from
typejoin, at least in 0.7.

It also makes `SVectors` of abstract types like `Number` become concrete
after numerical operations, which is inconsistent with base, but
probably not a problem for realisitc uses of this package.