Remove inferabilty of index_sizes helper test set.
This set was introduced to test an inference issue (https://github.com/JuliaLang/julia/issues/21244) in an old version of Julia.

It currently fails on Julia master because of https://github.com/JuliaLang/julia/issues/27276.

In my opinion, it's currently not really testing anything valuable, and the Julia PR that fixed the issue has associated test cases that should cover this.