Alternate approach to fix ##556
This restructures the similar method table along the lines that I suggested in https://github.com/JuliaArrays/StaticArrays.jl/pull/568#discussion_r241561280. I also used a bit of internals magic to match how the Base `OneTo` operates -- this is a little fragile but I think it is a good way to replicate how `OneTo` works.