Implement SA[...] syntax for SArray literals
Upside: This gives extremely attractive syntax for short static array
literals.

Downside: This is a syntax pun: the result of the `SA[...]` depends on
whether the name `SA` is bound to `StaticArrays.SA` or is local variable
of some other type.

However, this is no worse than Int[1,2,3], which is also a pun on the
array indexing syntax. Cases where this syntax could refer to something
other than StaticArrays.SA are detectable by the compiler and in
principle also accessible to a linter.