Improve inferrability of eigen()
Introduce `_make_static(::Eigen)` to allow the compiler to infer the
output type of eigen() as a Union of concrete types. This seems
necessary because the compiler's Union-splitting can't act on several
statements at once, so pulling out the values and vectors fields of the
Eigen individually will result in a pair of Unions which the compiler
considers uncorrellated. They are correllated however and the
_make_static function barrier allows the compiler to understand this.

Also return an Eigen from all the other utility functions for
convenience.