Add in-place matrix multiply-add (#738)
Add support for the new 5-argument in-place matrix multiplication in Julia v1.3.

For use in TrajectoryOptimization.jl normal SArrays work great, but we found that compilation times went through the roof when we tried larger problems. To my knowledge, SizedArray with 5-arg mul! is the only solution that works nearly as fast as SArrays at small sizes (thanks to loop unrolling!), scales well to large arrays (thanks to BLAS), and doesn't incur any memory allocations.