Fix accidential API breakage in one() (#763)
It turns out that Rotations relies on one(M) calling the constructor of
M, so transitioning to use of simliar_type in #690 broke Rotations.

It's arguably a bug to call the constructor rather than calling
similar_type, but let's avoid breaking compatibility for now.

(It's a bug because some sets of matrices defined by a type cannot
represent the multiplicative identity, for example the set of
skew-symmetric matrices. StaticArrays should likely assume that generic
operations like one() must devolve to a matrix type which can represent
anything in GL(n,R). For example, SMatrix.)