Add a work-around on Julia 1.0 for #857
`StaticVecOrMatLike` is a very large union, and this caused Julia 1.0.x to use a
large amount of memory when compiling StaticArrays in the presence of other
packages that also extend `LinearAlgebra.mul!` in a similar fashion. As an
example, this led to JuMP using more than 4 GB of RAM when running `using JuMP`,
causing its CI to crash.

Computing the eltypes within the function, rather than specifying them in the
type arguments greatly improves the issue, with `using JuMP` going from:
    julia> @time using JuMP
    [ Info: Precompiling JuMP [4076af6c-e467-56ae-b986-b466b2749572]
    Killed
to:
    julia> @time using JuMP
    [ Info: Precompiling JuMP [4076af6c-e467-56ae-b986-b466b2749572]
    97.868305 seconds (12.75 M allocations: 805.413 MiB, 0.30% gc time)