Code clean and reuse for `constructor`. (#1016)
* Define `construct_type`

Similar to `similar_type`, but not fallback to `SArray`. It is used to pick the most `concrete` constructor with `size`, `eltype` and `ndims` defined.

* Define general `constructor` with `construct_type`

* Define general `convert` with `construct_type`.

And fix for input with non-1 based axes.

* Make `FieldArray`'s general constructor based on `construct_type`

This also enable auto type promotion.

* `Constructor` and `convertor` clean.

With `construct_type`, there's no need to keep all these dispatches.
This also fix empty construction for `S/MVector`, and remove most of the ambiguities.

* Add more test

* Test fix and clean.

* Drop `StaticSquareMatrix` & add more test.

1. remove `StaticSquareMatrix` (`StaticMatrix{N,N}` should be shorter and clear enough)
2. Add missing Test.

* Add constructor test for `OffsetArray`.

* Rename `FirstClass` as `SizeEltypeAdaptable`

And convert comments to docstring.

* Replace `_NTuple` with `_TupleOf`

and typo fix.