make `@SVector` and `@SMatrix` calls default to `Float64` to match Base
This makes calls to `zeros`, `ones`, `rand`, and friend default to `Float64`
similar to Base. This can be a source of type instability when used
in the single argument version, and it is not transparent as to why.
Making `Float64` the default eliminates this source of error for the user.

Closes: https://github.com/JuliaArrays/StaticArrays.jl/issues/1164