fix: fix n-dimensional `view` into n-dimensional `SArray` with `Any` eltype (#1310)
As of the latest release, trying to `view` into an `SArray` with an
eltype of `Any` such that the view also returns an `SArray` would wrap
the returned array in a scalar. `_maybewrapscalar` assumes that if the
`eltype` of the array matches the type of the view, it must be a scalar
view. This assumption breaks down when the `eltype` is `Any`. MWE:

```julia
julia> A = SVector{4, Any}(1,2,3,4)

julia> I = SVector{2}(3, 1)

julia> view(A, I)
Scalar{Any}((Any[3, 1],))

julia> view(A, [3, 1])
2-element view(::SVector{4, Any}, [3, 1]) with eltype Any:
 3
 1
```

I've verified that all the added tests fail on the latest release and
pass with this PR.